Refactor wildcard IT into KustoSinkIT to reuse containers
- Fold shouldHandleWildcardSubscription into KustoSinkIT, sharing the
  existing Kafka/Connect/Proxy/Schema-Registry container stack instead
  of spinning up a second one (~280 lines of duplication removed,
  cuts IT runtime/Docker footprint roughly in half).
- Add produceJsonKafkaMessages(topic, vtype, n) helper so the wildcard
  scenario reuses performDataAssertions/getRecordsIngested.
- Allow deployConnector overrideProps to set 'topics.regex' (the
  default 'topics' key is dropped automatically when both are present,
  since Kafka Connect treats them as mutually exclusive) and to remove
  keys by setting them to null.
- Drop stray trailing blank line in .gitignore.
