+++ .gitignore
@@ -336,4 +336,3 @@ deps.txt
 .vscode/settings.json
 /kafka-sink-azure-kusto.ipr
 /kafka-sink-azure-kusto.iws
-
